orchestrator/requirements.txt
python-gvm
pyyaml
flask
lxml